Added gitignore file and work on books

+++ bookstorage/src/main/java/com/project/bookstorage/models/Book.java
@@ -13,7 +13,7 @@ import lombok.Data;
 public class Book {
     
     @Id
-    private int isbn;
+    private long isbn;
     private String title;
     private String author;
     private String genre;
