+++ bookstorage/src/main/java/com/project/bookstorage/services/BookService.java
@@ -5,8 +5,8 @@ import com.project.bookstorage.models.Book;
 
 public interface BookService {
     String addBook(Book book);
-    String deleteBook(int isbn);
+    String deleteBook(long isbn);
     String updateBook(Book book);
-    Book getBookById(int isbn);
+    Book getBookById(long isbn);
     BookResponse getAllBooks(int pageNo, int pageSize);
 }
